% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_box_height.R
\name{calc_box_height}
\alias{calc_box_height}
\title{Calculate height of box}
\usage{
calc_box_height(label)
}
\arguments{
\item{label}{The text to fit into the box, including line returns}
}
\value{
Numeric.  The height of the box
}
\description{
Calculate height of box
}
