% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_org_chart.R
\name{make_org_chart}
\alias{make_org_chart}
\title{Make Org Chart}
\usage{
make_org_chart(org_data)
}
\arguments{
\item{org_data}{A dataframe prepared by prepare_org_data()}
}
\value{
A grViz object suitable for plotting
}
\description{
Make Org Chart
}
