% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_edges_dtf.R
\name{oc_edges_dtf}
\alias{oc_edges_dtf}
\title{Create a dataframe of edges}
\usage{
oc_edges_dtf(.data, nodes)
}
\arguments{
\item{.data}{A dataframe of organisation data}

\item{nodes}{A daataframe of nodes calculated based on the organisation data}
}
\value{
A dataframe of edges for the organisation
}
\description{
Create a dataframe of edges
}
