% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_nodes_dtf.R
\name{oc_nodes_dtf}
\alias{oc_nodes_dtf}
\title{Create a dataframe of nodes}
\usage{
oc_nodes_dtf(.data, include_job_titles = TRUE)
}
\arguments{
\item{.data}{A dataframe of organisation info}

\item{include_job_titles}{Optionally exlude job titles, even if they are available}
}
\value{
A dataframe of nodes
}
\description{
Create a dataframe of nodes
}
