% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_root_nodes.R
\name{oc_root_nodes}
\alias{oc_root_nodes}
\title{Find the root nodes, which are nodes with no manager above them}
\usage{
oc_root_nodes(.data)
}
\arguments{
\item{.data}{A dataframe containing organisation details}
}
\value{
A character vector of names for the root nodes
}
\description{
Find the root nodes, which are nodes with no manager above them
}
